requirements.txt
markdown
flask
groq
python-dotenv
googlesearch-python